Bloc6/testgeodechet/requirements.txt
altair
pandas
streamlit
matplotlib
numpy
openpyxl
statsmodels
requests
boto3
shap
langchain
langchain-mistralai
python-dotenv
plotly